New column working days

+++ payroll/src/main/java/org/ideyatech/payroll/entity/UserCutOff.java
@@ -107,7 +107,7 @@ public class UserCutOff {
 		this.nontaxable = nontaxable;
 	}
 
-	public int getNumberofdependents() {
+	public double getNumberofdependents() {
 		return numberofdependents;
 	}
